Use basic implementations for editor and images
